fix(kernel): rollback process storage projection lease sets
Roll branch, owner-home, and Fleet-shared provider leases through one
retention-aware cleanup path. Failed provider teardown keeps a single
authenticated blocker for retry; successful rollback removes the
ephemeral mount root. Parent-token fault injection now clears only the
matching ordinal.

Tracking #1710
