+++ changes/1710.security.md
@@ -6,3 +6,7 @@
 - Kept user-private state excluded from process provider projections and
   retained revoked projection leases when cleanup fails; retained state blocks
   duplicate projection creation until cleanup succeeds. Tracking #1710.
+- Rolled back every issued branch, owner-home, and Fleet-shared lease when a
+  provider launch fails; retained the projection only when authenticated
+  cleanup cannot complete and removed the ephemeral mount root when rollback
+  succeeds. Tracking #1710.
